// src/app/flow-councils/types/recipient.ts
import { ProjectMetadata } from "@/types/project";

export type Recipient = {
  id: string;
  address: `0x${string}`;
  metadata: ProjectMetadata;
  bannerCid: string;
  twitter: string;
  flowRate: bigint;
  units: number;
  placeholderLogo: string;
  placeholderBanner: string;
};

export enum SortingMethod {
  RANDOM = "Random Order",
  ALPHABETICAL = "Alphabetical",
  POPULAR = "Popular",
}